==> src/00_libs.R <==

library(dplyr)
library(data.table)
library(tidyr)
library(highcharter)
library(quantmod)
library(lubridate)
library(corrplot)
library(DataExplorer)
library(stringr)

library(car)
library(caret)
library(forecast)
library(randomForest)
library(gbm)
library(xgboost)
library(lightgbm)
library(e1071)
library(Matrix)

library(DBI)
library(RMariaDB)
library(Ckmeans.1d.dp)
pwd = read.table("../db_password.txt", header = T)
